+++ src/debra.nim
@@ -21,6 +21,7 @@ import ./debra/typestates/neutralize
 import ./debra/typestates/advance
 import ./debra/typestates/slot
 import ./debra/managed
+import ./debra/convenience
 
 export types
 export signal.setGlobalManager, signal.installSignalHandler
@@ -36,6 +37,7 @@ export neutralize
 export advance
 export slot
 export managed
+export convenience
 
 proc registerThread*[MaxThreads: static int](
   manager: var DebraManager[MaxThreads]